requirements.txt
fastapi[all]
uvicorn[standard]
binance==0.3
flask
gunicorn
requests==2.31.0
bcrypt==4.0.1
watchdog==3.0.0
watchfiles==0.20.0
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
pymongo==4.4.1

python-binance==1.0.19
cryptocode==0.1
